Last commit before poop

+++ app/Http/Controllers/FavController.php
@@ -19,27 +19,43 @@ class FavController extends Controller
             $favoritePosts = $user->favoritePosts()->get();
             // Maintenant, $favoritePosts contient une collection de tous les articles favorisés par l'utilisateur
             return view('fav.mainFavorite', ['favoritePosts' => $favoritePosts, 'menus' => menu::all(), 'footers' => footer::all()]);
-        } else {
-            // L'utilisateur n'est pas authentifié, rediriger ou afficher un message d'erreur
+        }
+
+        else {
+
+            return redirect(route('auth.login'));
+
         }
     }
 
 
     public function toggleFavorite(Request $request)
     {
-        $postId = $request->post_id;
         $user = auth()->user();
-        $post = Post::findOrFail($postId);
 
-        if ($user->favoritePosts()->where('post_id', $postId)->exists()) {
-            $user->favoritePosts()->detach($postId);
-            return redirect()->route('blog.index')->with('fail', "L'article a bien été retiré des favoris");
+        if ($user) {
+
+
+            $postId = $request->post_id;
+            $user = auth()->user();
+            $post = Post::findOrFail($postId);
+
+            if ($user->favoritePosts()->where('post_id', $postId)->exists()) {
+                $user->favoritePosts()->detach($postId);
+                return redirect()->route('blog.index')->with('fail', "L'article a bien été retiré des favoris");
+            }
+
+            else {
+                $user->favoritePosts()->attach($postId);
+                return redirect()->route('blog.index')->with('success', "L'article a bien été mis en favoris");
+
+
+            }
+
         }
 
         else {
-            $user->favoritePosts()->attach($postId);
-            return redirect()->route('blog.index')->with('success', "L'article a bien été mis en favoris");
-
+            return redirect(route('auth.login'));
         }
     }
